requirements.txt
pyqt6>=6.6.0
PyQt6-sip>=13.6.0
qtawesome>=1.3.0
azure-eventhub>=5.11.0
requests>=2.31.0
faker>=20.0.0
python-dotenv>=1.0.0
